Add homepage and router, update minimum iOS version
A new homepage and router have been added to manage routes in the application. The homepage includes a basic scaffold with a counter and an increment function. The minimum supported iOS version has been updated to 12.0+, up from 11.0, to meet the latest standard requirements. Commit also includes updates to several dependencies.
