+++ lib/router.dart
@@ -1,0 +1,3 @@
+class MyRouter {
+  static const home = '/home';
+  }
